test=develop, bug fix for test_fleet_base
