+++ python/paddle/fluid/tests/unittests/test_fleet_base.py
@@ -151,109 +151,8 @@ class TestFleetBase(unittest.TestCase):
         optimizer = fleet.distributed_optimizer(optimizer, strategy=strategy)
 
     def test_exception(self):
-        from paddle.distributed.fleet import fleet
-        self.assertRaises(Exception, fleet.init_worker)
-
-    def _net(self):
-        import paddle
-
-        input_x = paddle.fluid.layers.data(
-            name="x", shape=[32], dtype='float32')
-        input_y = paddle.fluid.layers.data(name="y", shape=[1], dtype='int64')
-
-        fc_1 = paddle.fluid.layers.fc(input=input_x, size=64, act='tanh')
-        fc_2 = paddle.fluid.layers.fc(input=fc_1, size=64, act='tanh')
-        prediction = paddle.fluid.layers.fc(input=[fc_2], size=2, act='softmax')
-        cost = paddle.fluid.layers.cross_entropy(
-            input=prediction, label=input_y)
-        avg_cost = paddle.fluid.layers.mean(x=cost)
-        return avg_cost
-
-    def test_collective_minimize(self):
-        import paddle
         import paddle.distributed.fleet as fleet
-        import paddle.fluid.incubate.fleet.base.role_maker as role_maker
-
-        avg_cost = self._net()
-        role = role_maker.PaddleCloudRoleMaker(is_collective=True)
-        fleet.init(role)
-        strategy = fleet.DistributedStrategy()
-        optimizer = paddle.optimizer.SGD(learning_rate=0.001)
-        optimizer = fleet.distributed_optimizer(optimizer, strategy=strategy)
-        optimizer.minimize(avg_cost)
-
-    def test_ps_minimize(self):
-        import paddle
-        from paddle.distributed.fleet import fleet
-        import paddle.fluid.incubate.fleet.base.role_maker as role_maker
-
-        os.environ["TRAINING_ROLE"] = "PSERVER"
-        os.environ["POD_IP"] = "127.0.0.1"
-        os.environ["PADDLE_PORT"] = "36001"
-
-        avg_cost = self._net()
-        role = role_maker.PaddleCloudRoleMaker(is_collective=False)
-        fleet.init(role)
-        strategy = paddle.distributed.fleet.DistributedStrategy()
-        strategy.a_sync = False
-        optimizer = paddle.optimizer.SGD(learning_rate=0.001)
-        optimizer = fleet.distributed_optimizer(optimizer, strategy=strategy)
-        optimizer.minimize(avg_cost)
-
-        place = fluid.CPUPlace()
-        exe = fluid.Executor(place)
-        pe = fluid.ParallelExecutor(use_cuda=False, loss_name=avg_cost.name)
-        compiled_prog = fluid.compiler.CompiledProgram(
-            fluid.default_main_program())
-        self.assertRaises(
-            Exception,
-            fleet.save_inference_model,
-            dirname='/tmp/',
-            feeded_var_names=['x', 'y'],
-            target_vars=[avg_cost],
-            executor=pe)
-
-        self.assertRaises(
-            Exception,
-            fleet.save_inference_model,
-            dirname='/tmp/',
-            feeded_var_names=['x', 'y'],
-            target_vars=[avg_cost],
-            executor="exe")
-
-        self.assertRaises(
-            Exception,
-            fleet.save_inference_model,
-            dirname='/tmp/',
-            feeded_var_names=['x', 'y'],
-            target_vars=[avg_cost],
-            executor=exe,
-            main_program=compiled_prog)
-
-        self.assertRaises(
-            Exception,
-            fleet.save_persistables,
-            dirname='/tmp/',
-            feeded_var_names=['x', 'y'],
-            target_vars=[avg_cost],
-            executor=pe)
-
-        self.assertRaises(
-            Exception,
-            fleet.save_persistables,
-            dirname='/tmp/',
-            feeded_var_names=['x', 'y'],
-            target_vars=[avg_cost],
-            executor="exe")
-
-        self.assertRaises(
-            Exception,
-            fleet.save_persistables,
-            dirname='/tmp/',
-            feeded_var_names=['x', 'y'],
-            target_vars=[avg_cost],
-            executor=exe,
-            main_program=compiled_prog)
+        self.assertRaises(Exception, fleet.init_worker)
 
 
 if __name__ == "__main__":
